requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
fuzzywuzzy
python-Levenshtein
requests
python-dotenv